+++ docs/api/data.rst
@@ -1,5 +1,5 @@
 ukis_pysat.\data
-================
+====================
 
 .. automodule:: ukis_pysat.data
   :members:
