--- docs/api/download.rst
@@ -1,7 +1,0 @@
-ukis_pysat.\download
-====================
-
-.. automodule:: ukis_pysat.download
-  :members:
-  :undoc-members:
-  :show-inheritance:
